modification fonction distance

+++ point.c
@@ -38,13 +38,13 @@ signed int getY(point p)
   return p->y;
 }
 
-float distanceEntreDeuxPoints(point p1, point p2)
+double distanceEntreDeuxPoints(point p1, point p2)
 {
   double x1 = getX(p1);
   double x2 = getX(p2);
   double y1 = getY(p1);
   double y2 = getY(p2);
-  float tmp =  (x2 - x1) + (y2 - y1);
+  double tmp =  (x2 - x1) + (y2 - y1);
 	if (tmp<0)
 		tmp=-tmp;
   return tmp;
